refactor: move TaskRunner + AssessmentOrchestrator to @polpo-ai/core
TaskRunner:
- Replaced node:fs (readFileSync/existsSync) with ctx.readRunLog port
- Replaced node:path with pure string helpers (pathJoin/pathDirname)
- Replaced getSocketPath with ctx.notifySocketPath port
- Replaced validateProviderKeys with ctx.validateProviderKeys port

AssessmentOrchestrator:
- Replaced node:fs/node:path with AssessmentPorts (fileExists, readDir, etc.)
- Replaced queryOrchestratorText with ctx.queryLLM port
- Replaced generateAnswer with AssessmentPorts.generateAnswer
- Replaced classifyAsQuestion with AssessmentPorts.classifyAsQuestion
- Moved assessment-prompts.ts and question-detector.ts to core (pure logic)

Root files become thin re-export shims that inject Node.js implementations.
All 1153 tests pass, zero typecheck errors.
